moved test to cql test base
